services/media/service/src/domains/movies/localization/localizable-movie-genre-db-message-handlers.spec.ts
import { optional } from '@axinom/mosaic-db-common';
import { SubjectType } from '@axinom/mosaic-id-guard';
import { TokenResult } from '@axinom/mosaic-id-link-be';
import { MessageEnvelopeOverrides } from '@axinom/mosaic-message-bus';
import { MessagingSettings } from '@axinom/mosaic-message-bus-abstractions';
import {
  DeleteLocalizationSourceEntityCommand,
  LocalizationServiceMultiTenantMessagingSettings,
  UpsertLocalizationSourceEntityCommand,
} from '@axinom/mosaic-messages';
import {
  StoreOutboxMessage,
  TransactionalInboxMessage,
} from '@axinom/mosaic-transactional-inbox-outbox';
import { stub } from 'jest-auto-stub';
import 'jest-extended';
import { ClientBase } from 'pg';
import { PublicationConfig } from 'rascal';
import { Config } from '../../../common';
import { createTestConfig } from '../../../tests/test-utils';
import { LOCALIZATION_MOVIE_GENRE_TYPE } from './constants';
import {
  LocalizableMovieGenreCreatedDbMessageHandler,
  LocalizableMovieGenreDbEvent,
  LocalizableMovieGenreDeletedDbMessageHandler,
  LocalizableMovieGenreUpdatedDbMessageHandler,
} from './localizable-movie-genre-db-message-handlers';

const serviceAccountToken = 'SERVICE_ACCOUNT_TOKEN';
jest.mock('@axinom/mosaic-id-link-be', () => {
  const originalModule = jest.requireActual('@axinom/mosaic-id-link-be');
  return {
    __esModule: true,
    ...originalModule,
    getServiceAccountToken: jest.fn(() =>
      Promise.resolve<TokenResult>({
        accessToken: serviceAccountToken,
        expiresInSeconds: 600,
        tokenType: SubjectType.ManagedServiceAccount,
      }),
    ),
  };
});

describe('Localizable Movie Genre DB trigger events', () => {
  let messages: {
    payload:
      | DeleteLocalizationSourceEntityCommand
      | UpsertLocalizationSourceEntityCommand;
    settings: MessagingSettings;
    overrides: MessageEnvelopeOverrides | undefined;
    options: PublicationConfig | undefined;
  }[] = [];
  let storeOutboxMessage: StoreOutboxMessage;
  let config: Config;

  const createMessage = (
    payload: LocalizableMovieGenreDbEvent,
    messageContext?: unknown,
  ) =>
    stub<TransactionalInboxMessage<LocalizableMovieGenreDbEvent>>({
      payload,
      ...optional(messageContext, () => ({ metadata: { messageContext } })),
    });

  beforeAll(() => {
    storeOutboxMessage = jest.fn(
      async (
        _aggregateId,
        messagingSettings,
        message,
        _client,
        envelopeOverrides,
        options,
      ) => {
        messages.push({
          payload: message as
            | DeleteLocalizationSourceEntityCommand
            | UpsertLocalizationSourceEntityCommand,
          settings: messagingSettings,
          overrides: envelopeOverrides,
          options,
        });
        return Promise.resolve(stub<any>()); // TODO: Change any to TransactionalMessage when mosaic libs are updated
      },
    );
    config = createTestConfig();
  });

  afterEach(async () => {
    messages = [];
  });

  afterAll(async () => {
    jest.restoreAllMocks();
  });

  describe('LocalizableMovieGenreCreatedDbMessageHandler', () => {
    let handler: LocalizableMovieGenreCreatedDbMessageHandler;

    beforeAll(() => {
      handler = new LocalizableMovieGenreCreatedDbMessageHandler(
        storeOutboxMessage,
        config,
      );
    });

    it('movie genre is inserted -> upsert message data is sent', async () => {
      // Arrange
      const payload: LocalizableMovieGenreDbEvent = {
        id: 1,
        title: 'Test Title',
      };

      // Act
      await handler.handleMessage(createMessage(payload), stub<ClientBase>());

      // Assert
      expect(messages).toEqual([
        {
          payload: {
            entity_id: payload.id.toString(),
            entity_title: payload.title,
            entity_type: LOCALIZATION_MOVIE_GENRE_TYPE,
            fields: {
              title: payload.title,
            },
            image_id: undefined,
            service_id: config.serviceId,
          },
          settings:
            LocalizationServiceMultiTenantMessagingSettings.UpsertLocalizationSourceEntity,
          overrides: {
            auth_token: serviceAccountToken,
          },
          options: {
            routingKey: `ax-localization-service.${config.tenantId}.${config.environmentId}.localization_source_entity.upsert`,
          },
        },
      ]);
    });
  });

  describe('LocalizableMovieGenreUpdatedDbMessageHandler', () => {
    let handler: LocalizableMovieGenreUpdatedDbMessageHandler;

    beforeAll(() => {
      handler = new LocalizableMovieGenreUpdatedDbMessageHandler(
        storeOutboxMessage,
        config,
      );
    });

    it('movie genre is updated -> upsert message data is sent', async () => {
      // Arrange
      const payload: LocalizableMovieGenreDbEvent = {
        id: 1,
        title: 'Test Title',
      };

      // Act
      await handler.handleMessage(createMessage(payload), stub<ClientBase>());

      // Assert
      expect(messages).toEqual([
        {
          payload: {
            entity_id: payload.id.toString(),
            entity_title: payload.title,
            entity_type: LOCALIZATION_MOVIE_GENRE_TYPE,
            fields: {
              title: payload.title,
            },
            image_id: undefined,
            service_id: config.serviceId,
          },
          settings:
            LocalizationServiceMultiTenantMessagingSettings.UpsertLocalizationSourceEntity,
          overrides: {
            auth_token: serviceAccountToken,
          },
          options: {
            routingKey: `ax-localization-service.${config.tenantId}.${config.environmentId}.localization_source_entity.upsert`,
          },
        },
      ]);
    });

    it('movie genre is updated with context -> upsert message data is sent', async () => {
      // Arrange
      const payload: LocalizableMovieGenreDbEvent = {
        id: 1,
        title: 'Test Title',
      };
      const context = {
        potentialContextProperty: 'test',
      };

      // Act
      await handler.handleMessage(
        createMessage(payload, context),
        stub<ClientBase>(),
      );

      // Assert
      expect(messages).toEqual([
        {
          payload: {
            entity_id: payload.id.toString(),
            entity_title: payload.title,
            entity_type: LOCALIZATION_MOVIE_GENRE_TYPE,
            fields: {
              title: payload.title,
            },
            image_id: undefined,
            service_id: config.serviceId,
          },
          settings:
            LocalizationServiceMultiTenantMessagingSettings.UpsertLocalizationSourceEntity,
          overrides: {
            auth_token: serviceAccountToken,
            message_context: context,
          },
          options: {
            routingKey: `ax-localization-service.${config.tenantId}.${config.environmentId}.localization_source_entity.upsert`,
          },
        },
      ]);
    });
  });

  describe('LocalizableMovieGenreDeletedDbMessageHandler', () => {
    let handler: LocalizableMovieGenreDeletedDbMessageHandler;

    beforeAll(() => {
      handler = new LocalizableMovieGenreDeletedDbMessageHandler(
        storeOutboxMessage,
        config,
      );
    });

    it('movie genre is deleted -> delete message data is sent', async () => {
      // Arrange
      const payload: LocalizableMovieGenreDbEvent = {
        id: 1,
      };

      // Act
      await handler.handleMessage(createMessage(payload), stub<ClientBase>());

      // Assert
      expect(messages).toEqual([
        {
          payload: {
            entity_id: payload.id.toString(),
            entity_type: LOCALIZATION_MOVIE_GENRE_TYPE,
            service_id: config.serviceId,
          },
          settings:
            LocalizationServiceMultiTenantMessagingSettings.DeleteLocalizationSourceEntity,
          overrides: {
            auth_token: serviceAccountToken,
          },
          options: {
            routingKey: `ax-localization-service.${config.tenantId}.${config.environmentId}.localization_source_entity.delete`,
          },
        },
      ]);
    });
  });
});
